app/requirements.txt
tensorflow==2.12.0
numpy==1.23.5
Flask>=2.0.1
Pillow>=8.3.2